# perf: Speedup `db.{commit|rollback|begin}` by ~1.4x