Refer: https://peps.python.org/pep-0249/#commit

Why is this faster?
- Typical queries are sent over query interface and reading the response
  which is "empty" is waste of time.
- conn.commit|rollback|begin send it as command just read "OK" packet
  and proceed. It has much lower overhead.

| Benchmark                                      | before | after                |
|------------------------------------------------|:------:|:--------------------:|
| bench_database_bench_empty_transaction_cycling | 234 us | 163 us: 1.43x faster |

Side effects:
1. No more "logging" of these queries, can be solved by just fake logging.
2. Things that don't start transaction will now also won't rollback.
   So `ping` might not even connect to DB. We can do it explicitily if we want to.