@ankush 2024-12-23
err...

- most test failures in mariadb are because they are chaining a read only transaction
- postgres failure is booking it's strict about existence of a transaction to "chain".